ChatBot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1

langchain==0.3.7
langchain-openai==0.2.5
langchain-community==0.3.5
faiss-cpu==1.8.0.post1

httpx==0.27.2
psycopg2-binary==2.9.9
